requirements.txt
fastapi
uvicorn
pytest
requests
loguru
pydantic
tensorflow
pandas
numpy
python-multipart
Pillow
cv2